Fix trash

+++ backend/apps/secret/routers.py
@@ -27,6 +27,7 @@ from toolkits.utils import check_password_complexity
 from apps.workspace.models import Workspace, Share
 from apps.config.schema import PasswordPolicySchema
 from toolkits.workspace import WorkspaceUtils
+from toolkits.secret import SecretUtils
 from fastapi.exceptions import HTTPException
 from apps.auth.tools import get_current_user
 from mongoengine.queryset.visitor import Q
@@ -34,6 +35,7 @@ from apps.auth.schema import LoggedUser
 from toolkits.crypto import CryptoUtils
 from fastapi.responses import Response
 from apps.folder.models import Folder
+from apps.workspace.models import Workspace
 from .models import Login, Secret
 from datetime import date, datetime
 from settings import settings
@@ -139,7 +141,8 @@ async def get_secret(
         secret = Secret.objects(pk=secret_id).get()
         secret_schema = secret.schema()
 
-        workspace, sym_key = WorkspaceUtils.get_workspace(secret.folder.workspace.pk, user)
+        workspace_pk = secret.folder.workspace.pk if secret.folder is not None else secret.trash.workspace.pk
+        workspace, sym_key = WorkspaceUtils.get_workspace(workspace_pk, user)
 
         decrypted_sym_key = CryptoUtils.rsa_decrypt(
             sym_key,
@@ -153,11 +156,13 @@ async def get_secret(
             get_protected_fields=True
         )
 
+        action = f"in folder {secret.folder.name}" if secret.folder is not None else "in trash"
+
         create_history(
             user=user.in_db.email,
             workspace=workspace.name,
             workspace_owner=workspace.owner.email,
-            action=f"Retreive secret for secret {decrypted_secret.name.value} in folder {secret.folder.name}"
+            action=f"Retreive secret for secret {decrypted_secret.name.value} {action}"
         )
 
         if user.email != decrypted_secret.created_by:
@@ -192,12 +197,6 @@ async def create_secret(
 ) -> str:
     try:
         folder: Folder = Folder.objects(pk=schema.secret.folder).get()
-        if folder.in_trash or folder.is_trash:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Can't create a secret inside the trash"
-            )
-
         workspace, sym_key = WorkspaceUtils.get_workspace(folder.workspace.pk, user)
 
         # Check if user is allowed to create a key in this workspace
@@ -249,11 +248,12 @@ async def move_key(
 ):
     try:
         secret: Secret = Secret.objects(pk=secret_id).get()
+        workspace, _ = WorkspaceUtils.get_workspace(secret.folder.workspace.pk if secret.trash is None else secret.trash.workspace.pk, user)
+        WorkspaceUtils.have_rights(workspace, user)
         new_folder: Folder = Folder.objects(pk=folder_id).get()
 
-        workspace, _ = WorkspaceUtils.get_workspace(secret.folder.workspace.pk, user)
-        WorkspaceUtils.have_rights(workspace, user)
-
+        if secret.trash is not None:
+            secret.trash = None
         secret.folder = new_folder
         secret.save()
 
@@ -283,10 +283,10 @@ async def update_secret(
 ) -> None:
     try:
         secret: Secret = Secret.objects(pk=secret_id).get()
-        if secret.folder.in_trash or secret.folder.is_trash:
+        if secret.folder is None:
             raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Can't edit a secret inside the trash"
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Secret is in trash"
             )
         workspace, sym_key = WorkspaceUtils.get_workspace(secret.folder.workspace.pk, user)
 
@@ -358,7 +358,13 @@ async def delete_secret(
 ) -> None:
     try:
         secret: Secret = Secret.objects(pk=secret_id).get()
-        workspace: Workspace = secret.folder.workspace
+
+        if secret.folder is not None:
+            raise HTTPException(
+                status_code= status.HTTP_400_BAD_REQUEST,
+                detail="You have to first put the secret in the trash to delete it"
+            )
+        workspace: Workspace = secret.trash.workspace
         WorkspaceUtils.have_rights(workspace, user)
         secret.delete()
 
@@ -366,7 +372,7 @@ async def delete_secret(
             user=user.email,
             workspace=workspace.name,
             workspace_owner=workspace.owner.email,
-            action=f"Delete secret {secret.name.value} in folder {secret.folder.name}"
+            action=f"Delete secret {secret.name.value} in trash"
         )
 
         logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} delete secret {secret.name.value}")
@@ -376,3 +382,81 @@ async def delete_secret(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Secret not found"
         )
+
+
+
+@router.delete(
+    path="/{secret_id}/trash",
+    summary="Move a secret to the trash",
+    status_code=status.HTTP_204_NO_CONTENT
+)
+async def move_to_trash_secret(
+    secret_id: str,
+    user: LoggedUser = Depends(get_current_user)
+) -> None:
+    try:
+        secret: Secret = Secret.objects(pk=secret_id).get()
+        if secret.trash is not None:
+            raise HTTPException(
+                status_code = status.HTTP_400_BAD_REQUEST,
+                detail = "The secret is alreay in the trash"
+            )
+
+        workspace : Workspace = secret.folder.workspace
+        WorkspaceUtils.have_rights(workspace, user)
+        trash : Trash = WorkspaceUtils.get_trash_folder(workspace)
+
+        SecretUtils.move_to_trash(secret, trash)
+
+        create_history(
+            user=user.email,
+            workspace=workspace.name,
+            workspace_owner=workspace.owner.email,
+            action=f"Move secret {secret.name.value} in the trash"
+        )
+
+        logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} move secret {secret.name.value} to trash")
+    
+    except Secret.DoesNotExist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Secret not found"
+        )
+
+@router.patch(
+    path="/{secret_id}/restore",
+    summary = "Restore a secret from the trash",
+    status_code = status.HTTP_204_NO_CONTENT
+)
+async def restore(
+    secret_id: str,
+    user : LoggedUser = Depends(get_current_user)
+) -> None:
+    try:
+        secret: Secret = Secret.objects(pk=secret_id).get()
+    except Secret.DoesNotExist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Secret not found"
+        )
+
+    if secret.trash is None:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail = "The secret is not in the trash"    
+        )
+
+    workspace : Workspace = secret.trash.workspace
+    WorkspaceUtils.have_rights(workspace, user)
+    SecretUtils.restore(secret, workspace)
+
+    create_history(
+        user=user.email,
+        workspace=workspace.name,
+        workspace_owner=workspace.owner.email,
+        action=f"Restored secret {secret.name.value}"
+    )
+
+    logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} restored secret {secret.name.value}")
+
+
